pyproject.toml
[tool.poetry]
name = "mvp1"
version = "0.1.0"
description = ""
authors = ["Luis Ramos <luis.ramost@ab-inbev.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.1"
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
